+++ lib/construct/index.js
@@ -5,8 +5,7 @@ const EventEmitter = require('events').EventEmitter;
 const intersect = require('lodash.intersection');
 const assign = require('object-assign');
 const Progress = require('./progress');
-const { promisify } = require('util')
-const through = require('through2');
+const Packer = require('./packer');
 const uuid = require('node-uuid');
 const crypto = require('crypto');
 const mkdirp = require('mkdirp');
@@ -90,7 +89,9 @@ class Constructor extends EventEmitter {
       this.purge.bind(this),
       this.timeout
     );
+
     this.emits = emits;
+    this.packer = new Packer();
   }
 
   /**
@@ -505,7 +506,7 @@ class Constructor extends EventEmitter {
     app.contextLog.info('Begin npm install & tarball repack', spec.name, spec);
     // MAKE ASYNC/AWAIT (await x 4)
     async.series({
-      unpack: this.unpack.bind(this, { content, installPath, statusWriter }),
+      unpack: this.packer.unpack.bind(this, { content, installPath, statusWriter }),
       install: this.install.bind(this, spec, installPath, statusWriter),
       pack: this.pack.bind(this, pkgDir, tarball, statusWriter),
       upload: this.upload.bind(this, spec, tarball, statusWriter)
@@ -537,31 +538,6 @@ class Constructor extends EventEmitter {
       .pipe(fs.createWriteStream(target))
       .once('error', done)
       .once('finish', done);
-  }
-
-  /**
-   * Unpack the base64 string content into a proper directory of code
-   *
-   * @param {Object} opts Options for process
-   * @param {Function} next Completion callback.
-   * @api public
-   */
-  unpack(opts, next) {
-    const { content, installPath, statusWriter } = opts;
-    const stream = through();
-    const statusKey = 'unpacking';
-    const done = once(statusWriter.writeWrap(statusKey, next));
-    statusWriter.writeStart(statusKey);
-
-    // MAKE ASYNC/AWAIT (wrap as promise)
-    stream
-      .pipe(zlib.Unzip()) // eslint-disable-line new-cap
-      .once('error', done)
-      .pipe(tar.extract(installPath))
-      .once('error', done)
-      .once('finish', done);
-
-    stream.end(Buffer.from(content, 'base64'));
   }
 
   /**
